started implementation of pure draw2d renderers to remove holongate plugins
git-svn-id: https://svn.jfire.org/svn/main/trunk/JFireScriptingEditor2D@13432 18c61ebf-f106-0410-9e1e-81fb82415b8f

+++ src/org/nightlabs/jfire/scripting/editor2d/impl/BarcodeDrawComponentImpl.java
@@ -231,7 +231,7 @@ implements BarcodeDrawComponent
 		getBarcode().setDrawingText(isHumanReadable());
 		getBarcode().setResolution(getModelResolution());
 		double barWidth = getBarWidth(getWidthScale());
-		getBarcode().setBarWidth(barWidth);
+		getBarcode().setBarWidth((int)barWidth);
 		getBarcode().setBarHeight(barcodeHeight);
 		if (logger.isDebugEnabled())
 		{
